requirements.txt
numpy==1.19.0
opencv-python==4.2.0.34
Pillow==7.1.2
pkg-resources==0.0.0
